- rename: database/model 폴더 이름 entity 로 변경
