+++ src/database/entity/device.entity.ts
@@ -1,4 +1,4 @@
-import { UserEntity } from "@src/database/model/user.entity";
+import { UserEntity } from "@src/database/entity/user.entity";
 
 export class DeviceEntity {
   pk?: string;
