+++ src/database/repository/bank.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from "@nestjs/common";
 import { DatabaseService } from "@src/database/database.service";
-import { BankEntity } from "@src/database/model/bank.entity";
+import { BankEntity } from "@src/database/entity/bank.entity";
 import { bank } from "../../../drizzle/schema/bank";
 
 @Injectable()
