+++ src/database/repository/route.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { DatabaseService } from "@src/database/database.service";
 import { stops } from "../../../drizzle/schema/stops";
-import { RouteEntity } from "@src/database/model/route.entity";
+import { RouteEntity } from "@src/database/entity/route.entity";
 import { route } from "../../../drizzle/schema/route";
 import { eq } from "drizzle-orm";
 import { alias } from "drizzle-orm/pg-core";
