+++ src/database/repository/user-bank.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { DatabaseService } from "@src/database/database.service";
 import { userBank } from "../../../drizzle/schema/user-bank";
-import { UserBankEntity } from "@src/database/model/user-bank.entity";
+import { UserBankEntity } from "@src/database/entity/user-bank.entity";
 import { eq } from "drizzle-orm";
 import { TxType } from "@src/global/types/tx.types";
