+++ src/discovery/discovery.service.ts
@@ -6,7 +6,7 @@ import { PaginationDto } from "@src/global/dto/pagination.dto";
 import { PotDto } from "@src/discovery/dto/pot.dto";
 import { PotSearchDto } from "@src/discovery/dto/pot-search.dto";
 import { PotRoomRepository } from "@src/database/repository/pot-room.repository";
-import { PotRoomEntity } from "@src/database/model/pot-room.entity";
+import { PotRoomEntity } from "@src/database/entity/pot-room.entity";
 
 @Injectable()
 export class PotGDiscoveryService {
