+++ src/pot/pot.service.ts
@@ -6,13 +6,13 @@ import { PotEventReducer } from "@src/pot/event/pot-event-reducer";
 import { PotCreateEventV1 } from "@src/pot/event/pot-create-event";
 import { RouteService } from "@src/discovery/route.service";
 import { DatabaseService } from "@src/database/database.service";
-import { RouteEntity } from "@src/database/model/route.entity";
+import { RouteEntity } from "@src/database/entity/route.entity";
 import { randomUUID } from "node:crypto";
-import { PotRoomEntity } from "@src/database/model/pot-room.entity";
+import { PotRoomEntity } from "@src/database/entity/pot-room.entity";
 import { TxType } from "@src/global/types/tx.types";
 import { PotEventRepository } from "@src/database/repository/pot-event.repository";
 import { PotRoomRepository } from "@src/database/repository/pot-room.repository";
-import { UserPotRoomEntity } from "@src/database/model/user-pot-room.entity";
+import { UserPotRoomEntity } from "@src/database/entity/user-pot-room.entity";
 import { UserPotRoomRepository } from "@src/database/repository/user-pot-room.repository";
 import { BaseResultDto } from "@src/global/dto/base-result.dto";
 import { Pot } from "@src/pot/model/pot";
